+++ bexpr.go
@@ -32,6 +32,7 @@ type Evaluator struct {
 
 // CreateEvaluator is used to create and configure a new Evaluator, the expression
 // will be used by the evaluator when evaluating against any supplied datum.
+// By default the evaluator will error after 2 million expressions.
 // The following Option types are supported:
 // WithHookFn, WithMaxExpressions, WithTagName, WithUnknownValue.
 func CreateEvaluator(expression string, opts ...Option) (*Evaluator, error) {
@@ -39,6 +40,9 @@ func CreateEvaluator(expression string, opts ...Option) (*Evaluator, error) {
 	var parserOpts []grammar.Option
 	if parsedOpts.withMaxExpressions != 0 {
 		parserOpts = append(parserOpts, grammar.MaxExpressions(parsedOpts.withMaxExpressions))
+	} else {
+		// Use sane default as large expressions consume significant memory
+		parserOpts = append(parserOpts, grammar.MaxExpressions(2000000))
 	}
 
 	ast, err := grammar.Parse("", []byte(expression), parserOpts...)