+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Improvements
 
 ### Changes
+- Adds a default of 2 million for max evaluated expressions.
 
 ### Fixed